Move Entites related to cup phase to separate module
