+++ ippon/authorization_views.py
@@ -4,7 +4,7 @@ from rest_framework.response import Response
 
 import ippon.models
 import ippon.models.fight
-from ippon.models import Group, GroupPhase, CupPhase
+from ippon.models import Group, GroupPhase
 import ippon.models.team_fight as tfm
 import ippon.models.team as tem
 import ippon.models.tournament as tm
@@ -94,12 +94,6 @@ def group_authorization(request, pk, format=None):
 def group_phase_authorization(request, pk, format=None):
     group_phase = get_object_or_404(GroupPhase.objects.all(), pk=pk)
     return has_tournament_authorization([True, False], group_phase.tournament.id, request)
-
-
-@api_view(['GET'])
-def cup_phase_authorization(request, pk, format=None):
-    cup_phase = get_object_or_404(CupPhase.objects.all(), pk=pk)
-    return has_tournament_authorization([True, False], cup_phase.tournament.id, request)
 
 
 @api_view(['GET'])
