+++ ippon/cup_phase/cup_phase_views_test.py
